Successfully enable app_dlopen_to_so mechanism
But still needed to be tested.

Now I only tested one app which uses JNI to load the libjniLib.so,
and print the string returned by the function helloString() in libjniLib.so .
helloString() calls the function char* get_str() and get the hello string,
and the get_str() is in both libhooked.so and libhooking.so .
Therefore, I can test whether the hooking mechanism works by writing rules
in /system/nh_file.txt and seeing what is printed by the app.

Change-Id: I8b8c7ca15ae36c942adb012e848e3685b83aef71